Split code into modules
- Broke each grid type apart into its own separate module
- Created a `util` and `plot` module
- Added a separate constants module
- Removed unused imports
